requirements.txt
websocket-client
requests
enum